sprint 6 Исправление замечаний
